braveza

+++ src/engine/include/Camera.hpp
@@ -10,15 +10,14 @@ class Camera {
   Point position;
   Point lookAt;
   Point up;
-  int fov;
+  float fov;
   float nearP;
   float farP;
 
   Camera();
-  Camera(Point position, Point lookAt, Point up, int fov, float nearP,
+  Camera(Point position, Point lookAt, Point up, float fov, float nearP,
          float farP);
 
-  //   std::string toString();
 };
 
 #endif  // CAMERA_HPP
